update gat penn training script
